requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-ipware==4.0.2
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
